bigquery_data_cleanup/requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.13.0
gunicorn==21.2.0
google-cloud-bigquery-storage==2.24.0
